fix: delete provisioner modal
