+++ web/src/i18n/i18n-types.ts
@@ -2487,15 +2487,23 @@ type RootTranslation = {
 			 */
 			header: string
 			/**
-			 * P​r​o​v​i​s​i​o​n​i​n​g​ ​s​t​a​t​i​o​n​ ​d​o​c​k​e​r​ ​s​e​t​u​p​ ​c​o​m​m​a​n​d
-			 */
-			cardTitle: string
-			/**
 			 * I​n​ ​o​r​d​e​r​ ​t​o​ ​b​e​ ​a​b​l​e​ ​t​o​ ​p​r​o​v​i​s​i​o​n​ ​y​o​u​r​ ​Y​u​b​i​K​e​y​s​,​ ​f​i​r​s​t​ ​y​o​u​ ​n​e​e​d​ ​t​o​ ​s​e​t​ ​u​p​
 		​ ​ ​ ​ ​ ​ ​ ​ ​p​h​y​s​i​c​a​l​ ​m​a​c​h​i​n​e​ ​w​i​t​h​ ​U​S​B​ ​s​l​o​t​.​ ​R​u​n​ ​p​r​o​v​i​d​e​d​ ​c​o​m​m​a​n​d​ ​o​n​ ​y​o​u​r​ ​c​h​o​s​e​n​
 		​ ​ ​ ​ ​ ​ ​ ​ ​m​a​c​h​i​n​e​ ​t​o​ ​r​e​g​i​s​t​e​r​ ​i​t​ ​a​n​d​ ​s​t​a​r​t​ ​p​r​o​v​i​s​i​o​n​i​n​g​ ​y​o​u​r​ ​k​e​y​s​.
 			 */
 			content: string
+			dockerCard: {
+				/**
+				 * P​r​o​v​i​s​i​o​n​i​n​g​ ​s​t​a​t​i​o​n​ ​d​o​c​k​e​r​ ​s​e​t​u​p​ ​c​o​m​m​a​n​d
+				 */
+				title: string
+			}
+			tokenCard: {
+				/**
+				 * A​c​c​e​s​s​ ​t​o​k​e​n
+				 */
+				title: string
+			}
 		}
 		list: {
 			headers: {
@@ -2534,10 +2542,16 @@ type RootTranslation = {
 			}
 		}
 		messages: {
-			/**
-			 * C​o​m​m​a​n​d​ ​c​o​p​i​e​d​.
-			 */
-			codeCopied: string
+			copy: {
+				/**
+				 * T​o​k​e​n​ ​c​o​p​i​e​d
+				 */
+				token: string
+				/**
+				 * C​o​m​m​a​n​d​ ​c​o​p​i​e​d
+				 */
+				command: string
+			}
 		}
 	}
 	openidAllow: {
@@ -5764,15 +5778,23 @@ export type TranslationFunctions = {
 			 */
 			header: () => LocalizedString
 			/**
-			 * Provisioning station docker setup command
-			 */
-			cardTitle: () => LocalizedString
-			/**
 			 * In order to be able to provision your YubiKeys, first you need to set up
 		        physical machine with USB slot. Run provided command on your chosen
 		        machine to register it and start provisioning your keys.
 			 */
 			content: () => LocalizedString
+			dockerCard: {
+				/**
+				 * Provisioning station docker setup command
+				 */
+				title: () => LocalizedString
+			}
+			tokenCard: {
+				/**
+				 * Access token
+				 */
+				title: () => LocalizedString
+			}
 		}
 		list: {
 			headers: {
@@ -5811,10 +5833,16 @@ export type TranslationFunctions = {
 			}
 		}
 		messages: {
-			/**
-			 * Command copied.
-			 */
-			codeCopied: () => LocalizedString
+			copy: {
+				/**
+				 * Token copied
+				 */
+				token: () => LocalizedString
+				/**
+				 * Command copied
+				 */
+				command: () => LocalizedString
+			}
 		}
 	}
 	openidAllow: {
