+++ web/src/i18n/pl/index.ts
@@ -1083,10 +1083,15 @@ const pl: Translation = {
     noLicenseMessage: 'Nie masz licencji na tę funkcję.',
     provisioningStation: {
       header: 'Stacja provisionująca YubiKey',
-      cardTitle: 'Komenda uruchamiająca stację',
       content: `Aby móc sprovisionować YubiKeya, należy najpierw skonfigurować
         fizyczną maszynę z gniazdem USB. Uruchom podane polecenie na wybranej maszynie
         aby zarejestrować maszynę i rozpocząć generowanie kluczy.`,
+      tokenCard: {
+        title: 'Token autoryzacyjny',
+      },
+      dockerCard: {
+        title: 'Przykład Docker',
+      },
     },
     list: {
       headers: {
@@ -1104,7 +1109,10 @@ const pl: Translation = {
       },
     },
     messages: {
-      codeCopied: 'Komenda skopiowana.',
+      copy: {
+        command: 'Komenda skopiowa',
+        token: 'Token skopiowany',
+      },
     },
   },
   openidAllow: {
