+++ web/src/pages/provisioners/style.scss
@@ -156,91 +156,6 @@
       @include media-breakpoint-up(lg) {
         grid-row: 2;
       }
-
-      & > .provisioning-setup {
-        box-sizing: border-box;
-        padding: 1.5rem;
-
-        @include media-breakpoint-down(lg) {
-          margin: 0 2rem;
-          width: calc(100% - 4rem);
-        }
-
-        @include media-breakpoint-up(lg) {
-          padding: 4rem;
-          margin: 0 4rem 2.5rem;
-          width: calc(100% - 8rem);
-        }
-
-        @include media-breakpoint-up(xxl) {
-          margin: 0;
-          width: 100%;
-        }
-
-        & > h4 {
-          @include modal-header;
-
-          text-align: center;
-          margin-bottom: 1.5rem;
-          @include media-breakpoint-up(lg) {
-            margin-bottom: 2rem;
-          }
-        }
-
-        & > .image-row {
-          height: 75px;
-          display: flex;
-          flex-flow: row;
-          align-items: center;
-          justify-content: center;
-          align-content: center;
-        }
-
-        & > p {
-          word-break: normal;
-          text-align: center;
-          @include regular-text;
-
-          margin-bottom: 1.5rem;
-          max-width: 100%;
-          color: var(--gray-dark);
-          font-size: 12px;
-
-          @include media-breakpoint-up(lg) {
-            margin-bottom: 2.5rem;
-          }
-        }
-
-        & > .expandable-card {
-          width: 100%;
-          margin-top: 2rem;
-          @include media-breakpoint-up(lg) {
-            margin-top: 4rem;
-          }
-
-          & > .expanded-content {
-            overflow: hidden;
-
-            & > p {
-              max-width: 100%;
-              text-overflow: ellipsis;
-              text-align: left;
-              word-break: break-all;
-              white-space: normal;
-              @include small-text;
-
-              line-height: 22px;
-            }
-          }
-        }
-        .command-skeleton {
-          line-height: 1;
-          height: 198px;
-          width: 100%;
-          margin-top: 20px;
-          border-radius: 12px;
-        }
-      }
     }
 
     @include media-breakpoint-up(xxl) {
